refactor: 모달, 팝오버 컴포넌트 이동 및 수정 (mina-gwak/issue-tracker#28)
