+++ FE/src/components/IssueTable/PopOver/AuthorPopOver.tsx
@@ -1,11 +1,10 @@
 import * as S from '@components/IssueTable/PopOver/PopOver.style';
 import Image from '@components/common/Image';
 import { IMAGE_SIZE } from '@components/common/Image/constants';
-import PopOver from '@components/common/PopOver';
 
 const AuthorPopOver = () => {
   return (
-    <PopOver>
+    <S.Wrapper>
       <S.ContentsWrapper>
         <Image
           url='https://avatars.githubusercontent.com/u/62706988?v=4'
@@ -17,7 +16,7 @@ const AuthorPopOver = () => {
           <S.UserName>Jamie</S.UserName>
         </S.UserInformation>
       </S.ContentsWrapper>
-    </PopOver>
+    </S.Wrapper>
   )
 }
