--- FE/src/components/common/PopOver/index.tsx
@@ -1,11 +1,0 @@
-import * as S from '@components/common/PopOver/PopOver.style';
-
-export interface PopOverPropsType {
-  children?: JSX.Element;
-}
-
-const PopOver = ({ children }: PopOverPropsType) => {
-  return <S.Wrapper>{children}</S.Wrapper>;
-};
-
-export default PopOver;
